v2 Stage A move #9: gates policy + quality-gate retry-counter fix
Builds hooks/lib/policy/gates.mjs absorbing 4 gate hooks. Original logic
preserved as .legacy.mjs siblings.

Wrappers (4):
  mpl-finalize-gate    → handleFinalize     (orchestrates 4 child rules
                                              from contracts.mjs in-proc)
  mpl-quality-gate     → handleQuality
  mpl-ambiguity-gate   → handleAmbiguity
  mpl-phase-controller → handlePhaseTransition (Pass-A: simple cases)

CLOSES EVAL FINDING #6 — quality-gate retry-counter stuck:
  Before: catch swallows writeState error → rmSync score file
          unconditionally → adversarial_retry_count never advances →
          infinite retry loop.
  After:  rmSync ONLY after writeState succeeds. catch surfaces
          systemMessage ('could not persist retry state, will retry
          this round'); score file PRESERVED for next round so the
          counter actually advances. readFileSync catch treated as
          missing-file (fail-closed).
  Regression test asserts counter advances from 2→3 across a transient
  writeState throw, and escalation fires at maxRetries.

Pass-A scope note (phase-controller):
  Simple cases live in gates.mjs:
    mpl-init · mpl-decompose · mpl-ambiguity-resolve · phase1-plan
    phase1a-research · phase1b-plan · phase4-fix · phase5-finalize
    G4-hang · blocked_hook short-circuit
  Heavy cases still execute via .legacy.mjs runLegacyMain():
    phase2-sprint cohort lazy-init · release-gate · release-finalize
    phase3-gate · small-plan/sprint/verify
  Pass-B (lib/release/ migration) deferred.

Documented exception: handleFinalize cross-imports from contracts.mjs
(the only sanctioned L2↔L2 import — children that used to run as
spawnSync now run synchronously in-process; thrown errors demote to
'advisory' entries matching legacy fail-open).

Verify: 1,809 hooks tests pass (+39: policy-gates.test.mjs 39/39)
        · 0 regressions · 4 legacy gate test files all still pass
