+++ hooks/mpl-ambiguity-gate.mjs
@@ -1,68 +1,40 @@
 #!/usr/bin/env node
 /**
- * MPL Ambiguity Gate Hook (PreToolUse)
+ * MPL Ambiguity Gate Hook (PreToolUse Task|Agent → mpl-decomposer).
  *
- * Blocks Task(subagent_type="mpl-decomposer") if ambiguity_score is missing
- * or exceeds the threshold (0.2). Forces Stage 2 Ambiguity Resolution
- * (orchestrator drives mpl_score_ambiguity MCP tool loop inline) to
- * complete before decomposition can proceed.
+ * Thin stdin/stdout shim over `hooks/lib/policy/gates.mjs::handleAmbiguity`
+ * (Move #9). The policy module owns user_contract_set + goal_contract
+ * validity + ambiguity_score threshold check + override branch + the
+ * phase-reversion state mutations. The wrapper persists the returned
+ * stateMutations via writeState, emits stderr surfaces, and translates
+ * the decision back into the legacy { continue, reason } stdout contract.
  *
- * Matcher: Task|Agent (same as mpl-validate-output)
- * When ambiguity gate fails: continue=false (blocks the tool call)
- * When gate passes or not relevant: continue=true, suppressOutput=true
+ * Legacy stdout contract preserved:
+ *   allow / noop / bypass → { continue: true, suppressOutput: true }
+ *   block → { continue: false, reason: <verbose legacy reason string> }
+ *
+ * Original implementation: hooks/mpl-ambiguity-gate.legacy.mjs
  */
 
 import { dirname, join } from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
-import { existsSync, readFileSync } from 'fs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-const { isMplActive, readState, writeState } = await import(
+const { isMplActive, writeState } = await import(
   pathToFileURL(join(__dirname, 'lib', 'mpl-state.mjs')).href
 );
-const { readGoalContract } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-goal-contract.mjs')).href
-);
-
-/**
- * 0.16 Tier A' opt-out: legacy projects can disable the user-contract
- * gate via .mpl/config.json { "user_contract_required": false }. Default true.
- */
-function isUserContractRequired(cwd) {
-  try {
-    const cfgPath = join(cwd, '.mpl', 'config.json');
-    if (!existsSync(cfgPath)) return true;
-    const cfg = JSON.parse(readFileSync(cfgPath, 'utf-8'));
-    if (cfg && cfg.user_contract_required === false) return false;
-  } catch {
-    // fall through
-  }
-  return true;
-}
-
-/**
- * Goal Contract opt-out: legacy projects can disable the readiness gate via
- * .mpl/config.json { "goal_contract_required": false }. Default true.
- */
-function isGoalContractRequired(cwd) {
-  try {
-    const cfgPath = join(cwd, '.mpl', 'config.json');
-    if (!existsSync(cfgPath)) return true;
-    const cfg = JSON.parse(readFileSync(cfgPath, 'utf-8'));
-    if (cfg && cfg.goal_contract_required === false) return false;
-  } catch {
-    // fall through
-  }
-  return true;
-}
-
 const { readStdin } = await import(
   pathToFileURL(join(__dirname, 'lib', 'stdin.mjs')).href
 );
+const { handle: gatesHandle } = await import(
+  pathToFileURL(join(__dirname, 'lib', 'policy', 'gates.mjs')).href
+);
 
-const AMBIGUITY_THRESHOLD = 0.2;
+function pass() {
+  console.log(JSON.stringify({ continue: true, suppressOutput: true }));
+}
 
 async function main() {
   const input = await readStdin();
@@ -71,138 +43,51 @@ async function main() {
   try {
     data = JSON.parse(input);
   } catch {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-    return;
+    return pass();
   }
 
   const cwd = data.cwd || data.directory || process.cwd();
+  if (!isMplActive(cwd)) return pass();
 
-  if (!isMplActive(cwd)) {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-    return;
+  const toolName = data.tool_name || data.toolName || '';
+  const toolInput = data.tool_input || data.toolInput || {};
+
+  const decision = gatesHandle('ambiguity', {
+    cwd,
+    toolName,
+    toolInput,
+  });
+
+  // Persist any state mutations the policy module decided on (phase reverts,
+  // goal_contract syncs). This mirrors the legacy writeState calls.
+  if (decision.stateMutations && Object.keys(decision.stateMutations).length > 0) {
+    try {
+      writeState(cwd, decision.stateMutations);
+    } catch {
+      // Best-effort: never block the surfacing decision on disk failure
+      // (matches legacy behavior — writeState was unguarded but a throw
+      //  would have surfaced as an unhandled rejection caught by main()).
+    }
   }
 
-  // Only intercept Task/Agent calls targeting mpl-decomposer
-  const toolName = data.tool_name || '';
-  if (toolName !== 'Task' && toolName !== 'Agent') {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-    return;
+  if (decision.stderr) {
+    try { process.stderr.write(decision.stderr); } catch { /* ignore */ }
   }
 
-  const toolInput = data.tool_input || {};
-  const subagentType = toolInput.subagent_type || toolInput.subagentType || '';
-
-  if (subagentType !== 'mpl-decomposer' && subagentType !== 'mpl:mpl-decomposer') {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-    return;
-  }
-
-  // This IS a decomposer call — check ambiguity gate
-  const state = readState(cwd);
-  if (!state) {
-    console.log(JSON.stringify({
-      continue: false,
-      reason: '[MPL] ⛔ Decomposer BLOCKED: Cannot read MPL state. Ensure .mpl/state.json exists.'
-    }));
-    return;
-  }
-
-  // 0.16 Tier A': Step 1.5 User Contract Interview must complete before decomposition.
-  // Gate is additive to ambiguity score — BOTH must pass. Legacy projects that pre-date
-  // 0.16 can opt out via .mpl/config.json { "user_contract_required": false }.
-  const contractRequired = isUserContractRequired(cwd);
-  const contractSet = state.user_contract_set === true;
-  if (contractRequired && !contractSet) {
-    writeState(cwd, { current_phase: 'mpl-init' });
-    console.log(JSON.stringify({
-      continue: false,
-      reason: '[MPL] ⛔ Decomposer BLOCKED: user_contract_set is false. ' +
-        'Run Phase 0 Step 1.5 first: orchestrator inline loop calling mpl_classify_feature_scope MCP tool ' +
-        'to produce .mpl/requirements/user-contract.md, then mpl_state_write({user_contract_set:true}). ' +
-        'See commands/mpl-run-phase0.md Step 1.5. ' +
-        'To opt out in legacy projects: set user_contract_required=false in .mpl/config.json.'
-    }));
-    return;
-  }
-
-  // Goal Contract readiness: the ambiguity score is necessary but not
-  // sufficient. Before decomposition, freeze a contract that carries the
-  // goal/pivot/ontology/variation axes/evidence policy, then validate the disk
-  // artifact. This is the MPL counterpart to a Seed readiness gate.
-  if (isGoalContractRequired(cwd)) {
-    const goal = readGoalContract(cwd);
-    if (!goal.exists || !goal.valid) {
-      writeState(cwd, { current_phase: 'mpl-ambiguity-resolve' });
+  switch (decision.action) {
+    case 'noop':
+    case 'allow':
+    case 'bypass':
+      return pass();
+    case 'block':
       console.log(JSON.stringify({
         continue: false,
-        reason: '[MPL] ⛔ Decomposer BLOCKED: goal contract is missing or incomplete. ' +
-          `Write .mpl/goal-contract.yaml before decomposition. Missing: ${goal.missing.join(', ')}. ` +
-          'It must freeze source goal/user request, project pivot, ontology, variation axes, acceptance criteria, E2E policy, security policy, and completion evidence. ' +
-          'To opt out in legacy projects: set goal_contract_required=false in .mpl/config.json.'
+        reason: decision.reason,
       }));
       return;
-    }
-
-    if (state.goal_contract_set !== true || state.goal_contract_hash !== goal.contract.content_sha256) {
-      writeState(cwd, {
-        goal_contract_set: true,
-        goal_contract_path: goal.path,
-        goal_contract_hash: goal.contract.content_sha256,
-      });
-    }
+    default:
+      return pass();
   }
-
-  // Issue #51: override takes precedence over score. When the orchestrator
-  // records a user-halt (or SDK fallback) override, the gate must let the
-  // dispatch through without touching ambiguity_score. This keeps score
-  // truthful in state.json so downstream metrics/risk reports can surface
-  // residual ambiguity — we trade silence for honesty.
-  const override = state.ambiguity_override;
-  const overrideActive = override && override.active === true;
-
-  const score = state.ambiguity_score;
-  const hasScore = score !== null && score !== undefined;
-
-  if (overrideActive) {
-    // Record the bypass in stderr so it surfaces in transcripts/tool logs.
-    process.stderr.write(
-      `[MPL] Ambiguity gate bypassed by override (by="${override.by}", reason="${override.reason}", score=${hasScore ? score : 'null'})\n`,
-    );
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-    return;
-  }
-
-  if (!hasScore) {
-    // No score — block and revert phase. Router maps mpl-ambiguity-resolve
-    // back to mpl-run-phase0.md Step 1 Stage 2 so a stopped session resumes
-    // into the interview loop instead of becoming an orphan state.
-    writeState(cwd, { current_phase: 'mpl-ambiguity-resolve' });
-    console.log(JSON.stringify({
-      continue: false,
-      reason: '[MPL] ⛔ Decomposer BLOCKED: ambiguity_score not found in state. ' +
-        'Run Stage 2 first: call mpl_score_ambiguity MCP tool with pivot_points + user_responses and persist score via mpl_state_write. ' +
-        'Phase reverted to mpl-ambiguity-resolve. ' +
-        'If the interview should be halted without further questions, set ambiguity_override.active=true with a reason before retrying.'
-    }));
-    return;
-  }
-
-  if (score > AMBIGUITY_THRESHOLD) {
-    writeState(cwd, { current_phase: 'mpl-ambiguity-resolve' });
-    console.log(JSON.stringify({
-      continue: false,
-      reason: `[MPL] ⛔ Decomposer BLOCKED: ambiguity_score=${score} exceeds threshold ${AMBIGUITY_THRESHOLD}. ` +
-        'Run Stage 2 again: re-call mpl_score_ambiguity MCP tool with updated user_responses targeting the weakest dimension. ' +
-        'Phase reverted to mpl-ambiguity-resolve. ' +
-        'To halt the loop without passing the threshold, set ambiguity_override.active=true with a reason (preserves the true score for downstream reporting).'
-    }));
-    return;
-  }
-
-  // Gate passed
-  console.log(JSON.stringify({ continue: true, suppressOutput: true }));
 }
 
-main().catch(() => {
-  console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-});
+main().catch(() => pass());
