+++ hooks/mpl-phase-controller.mjs
@@ -1,302 +1,92 @@
 #!/usr/bin/env node
 /**
- * MPL Phase Controller Hook (Stop)
- * Manages phase transitions and loop continuation for the MPL pipeline.
+ * MPL Phase Controller Hook (Stop) — Move #9 thin wrapper.
  *
- * Based on: design doc section 9.2 hook 3 + OMC persistent-mode.mjs pattern
+ * Pass-A scope (per Move #9 plan): the policy module
+ * `hooks/lib/policy/gates.mjs::handlePhaseTransition` owns the decision
+ * skeleton for the simple phase cases (mpl-init, mpl-decompose,
+ * mpl-ambiguity-resolve, phase1-plan, phase1a-research, phase1b-plan,
+ * phase4-fix, phase5-finalize) plus the G4-hang short-circuit and the
+ * blocked_hook routing.
  *
- * Phase transitions:
- * - phase2-sprint: checks PLAN.md TODOs → all done → phase3-gate
- * - phase3-gate: checks gate results → all pass → phase5-finalize, any fail → phase4-fix
- * - phase4-fix: checks fix_loop_count → exceeded / stagnating → phase5-finalize, else continue
- * - phase5-finalize: completion message + deactivate MPL
+ * The wrapper still owns the heavy cases that touch the filesystem
+ * (phase2-sprint cohort lazy-init reads goal-contract; release-gate +
+ * release-finalize write release-manifest.json + evidence-summary.md +
+ * gate-results.json via atomicWriteFile; release-finalize also creates
+ * snapshot refs + user-visible artifacts; phase3-gate routes cohort-aware
+ * gate evidence + applies fix-loop accounting; small-plan + small-sprint +
+ * small-verify gate the small-pipeline against mvp_scope). Those file-write
+ * side effects migrate to a `lib/release/` module in Pass-B (NOT Move #9).
  *
- * Always returns continue: true to keep the pipeline loop running until completion.
+ * For Pass-A, the wrapper delegates ALL routing to the legacy main loop
+ * but exposes the same exports the existing test file relies on
+ * (checkPlanStatus, checkGateResults). The legacy file is preserved verbatim
+ * as mpl-phase-controller.legacy.mjs for emergency rollback.
+ *
+ * Decision flow:
+ *   1. Read stdin → state.
+ *   2. Run gates.handlePhaseTransition for the early-exit short-circuits
+ *      (G4-hang, blocked_hook, missing state) and simple phase cases.
+ *   3. If gates returns action='emit', persist state mutations + emit
+ *      the stopReason. Done.
+ *   4. If gates returns action='delegate-to-legacy', forward execution
+ *      to the legacy main loop which still owns the heavy cases.
  */
 
 import { dirname, join } from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
-import { existsSync, mkdirSync, readFileSync, writeFileSync, renameSync } from 'fs';
-import { randomBytes } from 'crypto';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
+const isMain = import.meta.url === pathToFileURL(process.argv[1] || '').href;
 
-// Import shared MPL state utility
-const { readState, writeState, isMplActive, checkConvergence } = await import(
+const { readState, writeState, isMplActive } = await import(
   pathToFileURL(join(__dirname, 'lib', 'mpl-state.mjs')).href
 );
-
-// Import shared stdin reader
 const { readStdin } = await import(
   pathToFileURL(join(__dirname, 'lib', 'stdin.mjs')).href
 );
-
-// Enforcement policy resolver (P0-2, #110)
-const { resolveRuleAction } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-enforcement.mjs')).href
-);
-
-// Stage A: goal-contract reader for the D-Q7 small-pipeline guard.
-const { readGoalContract } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-goal-contract.mjs')).href
-);
-
-// Stage A Phase 1.6c-ii: release-manifest serializers for the release-finalize
-// file-write step. Pure helpers; this hook owns the filesystem side.
-const {
-  buildReleaseManifest,
-  buildEvidenceSummary,
-  buildGateResultsSnapshot,
-  RELEASE_DIR_REL_PATH,
-} = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-release-manifest.mjs')).href
-);
-
-const { parsePhaseContractGraphText } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-phase-contract-graph.mjs')).href
-);
-const { parseDecompositionGoalTraceText } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-goal-trace.mjs')).href
-);
-
-// Exp22 R11 / #210 — codex r1 on PR #222. Phase-controller writes phase
-// transitions directly via writeState() which bypasses PreToolUse hooks,
-// so state-invariant I13 alone misses the controller's own transitions.
-// Call blockedPhaseTransitionReason() before any writeState that lands
-// in a phase listed in REQUIRES_PHASE0_ARTIFACTS.
-const { blockedPhaseTransitionReason } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-phase0-artifacts.mjs')).href
-);
-
-// #240 A3: ambiguity threshold + force-proceed config knobs.
 const { loadConfig } = await import(
   pathToFileURL(join(__dirname, 'lib', 'mpl-config.mjs')).href
 );
-
-// #240 A3: resolve the ambiguity threshold from .mpl/config.json with
-// the literal 0.2 as the default. Operators on exploration-mode
-// projects (research spikes, prototypes) can raise the threshold or
-// set `ambiguity.force_proceed_after_rounds: N` to allow
-// state.ambiguity_force_proceed = true to override after N rounds.
-function ambiguityConfig(cwd) {
-  const cfg = loadConfig(cwd);
-  const raw = cfg?.ambiguity?.threshold;
-  const threshold = typeof raw === 'number' && Number.isFinite(raw) ? raw : 0.2;
-  const fpRaw = cfg?.ambiguity?.force_proceed_after_rounds;
-  const forceProceedAfterRounds = Number.isInteger(fpRaw) && fpRaw > 0 ? fpRaw : null;
-  return { threshold, forceProceedAfterRounds };
-}
-
-function ambiguityRoundCount(state) {
-  const h = state?.ambiguity_history;
-  return Array.isArray(h) ? h.length : 0;
-}
-
-/**
- * Emit a stopReason and return true when the controller's intended
- * transition to `nextPhase` is blocked by missing Phase 0 artifacts.
- * Caller should `return` immediately after a truthy result to skip the
- * writeState. The transition is blocked, not deferred — the user must
- * run /mpl:mpl-resume or restore Phase 0 before the controller will
- * advance again.
- */
-function emitPhase0BlockIfNeeded(cwd, nextPhase) {
-  const reason = blockedPhaseTransitionReason(cwd, nextPhase);
-  if (!reason) return false;
-  console.log(JSON.stringify({
-    continue: true,
-    stopReason: reason,
-  }));
-  return true;
-}
-
-// Stage A Phase 1.6c-iii: snapshot ref + user-visible artifact creation.
-const { createSnapshotRef, attemptArtifactCreation } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-release-artifact.mjs')).href
+const { handle: gatesHandle } = await import(
+  pathToFileURL(join(__dirname, 'lib', 'policy', 'gates.mjs')).href
 );
 
-/**
- * Atomic file write — write to a sibling tmp file then rename onto the
- * final path. Removes the partial-file failure window claude #2 on PR
- * #187 flagged: if a write is interrupted mid-stream, the final path is
- * untouched (the tmp is orphaned, which a future release-finalize
- * re-run reclaims by writing a fresh tmp). All three release artifacts
- * use this so the on-disk state is always "all three present at the
- * same generation" or "the prior generation intact" — never a
- * partial-mix that a consumer could read between writes.
- */
-function atomicWriteFile(filePath, contents, mode) {
-  const tmp = `${filePath}.${randomBytes(4).toString('hex')}.tmp`;
-  writeFileSync(tmp, contents, { mode });
-  renameSync(tmp, filePath);
-}
-
-// G4 hang detection (#109)
-const { detectHang } = await import(
-  pathToFileURL(join(__dirname, 'lib', 'mpl-hang-detector.mjs')).href
+// Legacy module owns the heavy file-write paths (release-manifest /
+// snapshot ref / artifact attempt / atomicWriteFile) plus the exports the
+// existing test file imports (checkPlanStatus, checkGateResults). The
+// wrapper forwards stdin to the legacy main when gates.mjs returns
+// 'delegate-to-legacy'.
+const legacy = await import(
+  pathToFileURL(join(__dirname, 'mpl-phase-controller.legacy.mjs')).href
 );
 
+export const checkPlanStatus = legacy.checkPlanStatus;
+export const checkGateResults = legacy.checkGateResults;
+
+function emitContinueTrue() {
+  console.log(JSON.stringify({ continue: true, suppressOutput: true }));
+}
+
 /**
- * Check PLAN.md checkbox completion status
+ * Apply stop-hook emit envelope from a gates.handlePhaseTransition decision.
  */
-function checkPlanStatus(cwd) {
-  const planPaths = [
-    join(cwd, '.mpl', 'PLAN.md'),
-    join(cwd, 'PLAN.md'),
-  ];
-
-  for (const planPath of planPaths) {
-    if (!existsSync(planPath)) continue;
-
-    const content = readFileSync(planPath, 'utf-8');
-    // Tolerant regex: accepts whitespace variations, case-insensitive
-    // x/X, FAILED/failed. `~` marks a TODO as "deferred by decision"
-    // (#241 B6) — counts toward "no remaining work" for routing
-    // purposes so the operator does not have to forge `[x]` on a
-    // deliberately-deferred task.
-    const todoPattern = /###\s*\[\s*(x|X|FAILED|failed|~| )\s*\]/g;
-    const matches = [...content.matchAll(todoPattern)];
-
-    if (matches.length === 0) return { total: 0, completed: 0, failed: 0, deferred: 0 };
-
-    let completed = 0;
-    let failed = 0;
-    let deferred = 0;
-    for (const m of matches) {
-      const val = m[1].trim();
-      if (val.toLowerCase() === 'x') completed++;
-      else if (val.toLowerCase() === 'failed') failed++;
-      else if (val === '~') deferred++;
+function emitDecision(cwd, decision) {
+  if (decision.stateMutations && Object.keys(decision.stateMutations).length > 0) {
+    try {
+      writeState(cwd, decision.stateMutations);
+    } catch {
+      // Best-effort: never wedge the Stop hook on a disk error.
     }
-
-    return { total: matches.length, completed, failed, deferred };
   }
-
-  return null;
-}
-
-/**
- * Check gate results from state.
- *
- * Reads structured evidence (`hard1_baseline`, `hard2_coverage`, `hard3_resilience`)
- * written by `mpl-gate-recorder` from real Bash exit codes. Decision tree:
- *
- *   1. Any present structured entry with `exit_code !== 0` → anyFailed=true.
- *      Machine-recorded failure dominates self-reported legacy — never let legacy
- *      booleans mask a recorded nonzero exit (PR #119 review blocker).
- *   2. All 3 structured present and zero failures → allPassed=true.
- *   3. Partial structured (some present, none failed, some missing) → allPassed=false,
- *      anyFailed=false, source='structured'. Per issue #102 spec ("하나라도 missing →
- *      allPassed=false"), do NOT fall through to legacy. Once mpl-gate-recorder has
- *      produced any structured entry, the rest must follow.
- *   4. Zero structured + strict → blocked (no machine evidence at all).
- *   5. Zero structured + non-strict → legacy boolean fallback with caller-surfaced
- *      warn (transitional; retired when `enforcement.strict` rolls out, #110).
- *
- * @param {object} state - parsed `.mpl/state.json`
- * @param {{ strict?: boolean }} [opts] - strict mode disables zero-structured legacy fallback
- * @returns {{
- *   allPassed: boolean,
- *   anyFailed: boolean,
- *   source: 'structured' | 'legacy' | 'none',
- *   missingEvidence: string[],
- *   details: { hard1: boolean|null, hard2: boolean|null, hard3: boolean|null }
- * }}
- */
-function checkGateResults(state, opts = {}) {
-  const gates = (state && state.gate_results) || {};
-  const strict = opts.strict === true;
-
-  const structuredEntries = {
-    hard1: gates.hard1_baseline,
-    hard2: gates.hard2_coverage,
-    hard3: gates.hard3_resilience,
-  };
-
-  const isStructuredEntry = (e) => e && typeof e === 'object' && typeof e.exit_code === 'number';
-  const presentEntries = Object.values(structuredEntries).filter(isStructuredEntry);
-  const structuredCount = presentEntries.length;
-
-  const missingEvidence = [];
-  if (!isStructuredEntry(structuredEntries.hard1)) missingEvidence.push('hard1_baseline');
-  if (!isStructuredEntry(structuredEntries.hard2)) missingEvidence.push('hard2_coverage');
-  if (!isStructuredEntry(structuredEntries.hard3)) missingEvidence.push('hard3_resilience');
-
-  const detailsFromStructured = () => ({
-    hard1: isStructuredEntry(structuredEntries.hard1) ? structuredEntries.hard1.exit_code === 0 : null,
-    hard2: isStructuredEntry(structuredEntries.hard2) ? structuredEntries.hard2.exit_code === 0 : null,
-    hard3: isStructuredEntry(structuredEntries.hard3) ? structuredEntries.hard3.exit_code === 0 : null,
-  });
-
-  // Step 1: machine-recorded failure dominates. Even one structured entry with
-  // nonzero exit_code forces anyFailed=true, irrespective of legacy booleans.
-  // Closes PR #119 review blocker (legacy true + structured nonzero must not pass).
-  if (presentEntries.some(e => e.exit_code !== 0)) {
-    return {
-      allPassed: false,
-      anyFailed: true,
-      source: 'structured',
-      missingEvidence,
-      details: detailsFromStructured(),
-    };
+  const payload = { continue: decision.continue !== false };
+  if (decision.suppressOutput) {
+    payload.suppressOutput = true;
+  } else if (decision.stopReason) {
+    payload.stopReason = decision.stopReason;
   }
-
-  // Step 2: all 3 structured present and zero failures → genuine pass.
-  if (structuredCount === 3) {
-    return {
-      allPassed: true,
-      anyFailed: false,
-      source: 'structured',
-      missingEvidence: [],
-      details: { hard1: true, hard2: true, hard3: true },
-    };
-  }
-
-  // Step 3: partial structured (1 or 2 present, none failed, some missing).
-  // Issue #102 spec: "하나라도 missing → allPassed=false". Once gate-recorder has
-  // started producing structured entries, the rest are required — legacy fallback
-  // would let a phase-runner skip a gate by self-reporting only.
-  if (structuredCount > 0) {
-    return {
-      allPassed: false,
-      anyFailed: false,
-      source: 'structured',
-      missingEvidence,
-      details: detailsFromStructured(),
-    };
-  }
-
-  // Step 4: zero structured + strict → block (no machine evidence at all).
-  if (strict) {
-    return {
-      allPassed: false,
-      anyFailed: false,
-      source: 'structured',
-      missingEvidence,
-      details: detailsFromStructured(),
-    };
-  }
-
-  // Step 5: zero structured + non-strict → legacy boolean fallback (transitional).
-  // Phase3-gate caller surfaces a system-reminder ⚠ on `source === 'legacy'`.
-  // Retired once `enforcement.strict` ships (#110, P0-2).
-  const hardResults = [gates.hard1_passed, gates.hard2_passed, gates.hard3_passed];
-  const required = hardResults.filter(r => r !== null && r !== undefined);
-  const passed = required.filter(r => r === true);
-  const failed = required.filter(r => r === false);
-
-  return {
-    allPassed: failed.length === 0 && passed.length > 0,
-    anyFailed: failed.length > 0,
-    source: required.length === 0 ? 'none' : 'legacy',
-    missingEvidence,
-    details: {
-      hard1: gates.hard1_passed,
-      hard2: gates.hard2_passed,
-      hard3: gates.hard3_passed,
-    },
-  };
+  console.log(JSON.stringify(payload));
 }
 
 async function main() {
@@ -306,1083 +96,47 @@ async function main() {
   try {
     data = JSON.parse(input);
   } catch {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
+    emitContinueTrue();
     return;
   }
 
   const cwd = data.cwd || data.directory || process.cwd();
-
-  // Check if MPL is active
   if (!isMplActive(cwd)) {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
+    emitContinueTrue();
     return;
   }
 
-  let state = readState(cwd);
+  const state = readState(cwd);
   if (!state) {
-    console.log(JSON.stringify({ continue: true, suppressOutput: true }));
+    emitContinueTrue();
     return;
   }
 
-  // G4 (#109) — hang detection. Run before phase routing so a stalled
-  // session is marked even when the phase branch would otherwise return
-  // a generic "in progress" message.
-  //
-  // Three branches:
-  //   (a) Newly detected hang (state.last_tool_at is stale, no exempt status):
-  //       mark `session_status='verification_hang'` and emit the alarm banner.
-  //   (b) Already-marked verification_hang: short-circuit phase routing so
-  //       phase transitions (e.g. phase3-gate → phase5-finalize) cannot
-  //       advance silently while the session is awaiting user triage. The
-  //       resume skill clears the marker only after the user picks resume /
-  //       rollback / cancel — until then every Stop tick re-surfaces the
-  //       triage guidance. (PR #126 review #1 — was previously a fall-through.)
-  //   (c) paused_budget / paused_checkpoint: intentional pauses, untouched.
-  const hangDet = detectHang(state, Date.now());
-  if (hangDet.hung) {
-    try {
-      writeState(cwd, { session_status: 'verification_hang' });
-    } catch { /* best-effort marking — never fail Stop hook on disk error */ }
-    console.log(JSON.stringify({
-      continue: true,
-      stopReason: hangDet.reason,
-    }));
-    return;
+  // Pass-A: gates owns simple phase cases + G4 hang + blocked_hook short
+  // circuits. Heavy cases (phase2-sprint cohort init, release-gate,
+  // release-finalize, phase3-gate, small-*) return 'delegate-to-legacy'.
+  const config = loadConfig(cwd) || {};
+  let decision;
+  try {
+    decision = gatesHandle('phase_transition', { cwd, state, config });
+  } catch {
+    // Policy module crashed — fall through to legacy.
+    decision = { action: 'delegate-to-legacy' };
   }
-  if (state.session_status === 'verification_hang') {
-    console.log(JSON.stringify({
-      continue: true,
-      stopReason: '[MPL G4] Session is currently marked verification_hang. Phase routing is paused until user triage. Run /mpl:mpl-resume to choose: resume current phase, roll back, or cancel.',
-    }));
-    return;
-  }
-  if (state.session_status === 'blocked_hook') {
-    const hook = state.blocked_by_hook || 'unknown hook';
-    const blockedPhase = state.blocked_phase || state.current_phase || 'unknown phase';
-    const instruction = state.resume_instruction || 'Resolve the recorded hook block, then retry the transition.';
-    console.log(JSON.stringify({
-      continue: true,
-      stopReason:
-        `[MPL] Phase routing is paused by ${hook} for ${blockedPhase}. ` +
-        `${instruction} Run /mpl:mpl-resume to continue once the missing evidence is restored.`,
-    }));
+
+  if (decision && decision.action === 'emit') {
+    emitDecision(cwd, decision);
     return;
   }
 
-  const phase = state.current_phase;
-
-  switch (phase) {
-    case 'mpl-init': {
-      // Init phase: triage + interview in progress
-      console.log(JSON.stringify({
-        continue: true,
-        stopReason: '[MPL] Initialization in progress. Complete Triage → Stage 1 (PP Interview) → Stage 2 (Ambiguity Resolution) before Decomposition.'
-      }));
-      break;
-    }
-
-    case 'mpl-decompose': {
-      // GATE: ambiguity_score MUST exist AND meet threshold before decomposition proceeds.
-      // Stage 2 (Ambiguity Resolution) writes ambiguity_score to state via the mpl_score_ambiguity MCP tool.
-      // #240 A3: threshold + force-proceed are config-driven.
-      const ambiguityScore = state.ambiguity_score;
-      const hasScore = ambiguityScore !== null && ambiguityScore !== undefined;
-      const { threshold, forceProceedAfterRounds } = ambiguityConfig(cwd);
-
-      if (!hasScore) {
-        // Block decomposition — ambiguity resolution was skipped
-        writeState(cwd, { current_phase: 'mpl-ambiguity-resolve' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL] ⛔ Decomposition BLOCKED: ambiguity_score not found in state. ' +
-            'Reverting to Stage 2 Ambiguity Resolution. ' +
-            'Call mpl_score_ambiguity MCP tool with pivot_points + user_responses, then persist the result via mpl_state_write.'
-        }));
-      } else if (ambiguityScore > threshold) {
-        // #240 A3: force-proceed override — operator on exploration-mode
-        // projects can set ambiguity.force_proceed_after_rounds: N in
-        // config + state.ambiguity_force_proceed: true after N rounds
-        // to bypass threshold loop. Defaults to null (no override path).
-        const rounds = ambiguityRoundCount(state);
-        const forceProceed = state?.ambiguity_force_proceed === true &&
-          forceProceedAfterRounds !== null &&
-          rounds >= forceProceedAfterRounds;
-        if (forceProceed) {
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `[MPL] Decomposition: ambiguity_score=${ambiguityScore} exceeds threshold ${threshold}, but force-proceed override active (rounds=${rounds} >= ${forceProceedAfterRounds}). Proceeding with elevated ambiguity.`
-          }));
-        } else {
-          // Block decomposition — score exceeds threshold, re-run ambiguity resolution
-          writeState(cwd, { current_phase: 'mpl-ambiguity-resolve' });
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `[MPL] ⛔ Decomposition BLOCKED: ambiguity_score=${ambiguityScore} exceeds threshold ${threshold}. ` +
-              'Reverting to Stage 2 Ambiguity Resolution for additional Socratic resolution. ' +
-              'Re-call mpl_score_ambiguity MCP tool with updated user_responses targeting the weakest dimension.'
-          }));
-        }
-      } else {
-        // Score exists and meets threshold — proceed
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Decomposition: ambiguity_score=${ambiguityScore} (threshold: <=${threshold}). ✓ Proceed with micro-phase decomposition.`
-        }));
-      }
-      break;
-    }
-
-    case 'mpl-ambiguity-resolve': {
-      // Stage 2: Ambiguity Resolution in progress
-      const currentScore = state.ambiguity_score;
-      const hasCurrentScore = currentScore !== null && currentScore !== undefined;
-      // #240 A3: same config-driven threshold as the decompose case.
-      const { threshold: ambThreshold, forceProceedAfterRounds: ambForceProceedAfterRounds } = ambiguityConfig(cwd);
-
-      // #240 A3: same force-proceed override.
-      const ambRounds = ambiguityRoundCount(state);
-      const ambForceProceed = state?.ambiguity_force_proceed === true &&
-        ambForceProceedAfterRounds !== null &&
-        ambRounds >= ambForceProceedAfterRounds;
-      if (ambForceProceed && hasCurrentScore && currentScore > ambThreshold) {
-        writeState(cwd, { current_phase: 'mpl-decompose' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Ambiguity force-proceed (score=${currentScore}, rounds=${ambRounds} >= ${ambForceProceedAfterRounds}). Transitioning to Decomposition.`
-        }));
-        break;
-      }
-
-      if (hasCurrentScore && currentScore <= ambThreshold) {
-        // Threshold met — transition to decompose
-        writeState(cwd, { current_phase: 'mpl-decompose' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Ambiguity resolved: score=${currentScore} (<=${ambThreshold}). Transitioning to Decomposition.`
-        }));
-      } else {
-        const scoreInfo = hasCurrentScore ? ` Current score: ${currentScore}.` : '';
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Stage 2: Ambiguity Resolution in progress.${scoreInfo} Target: <=${ambThreshold}. ` +
-            'Drive the Socratic loop inline: call mpl_score_ambiguity MCP tool after each user response and persist via mpl_state_write.'
-        }));
-      }
-      break;
-    }
-
-    case 'phase1-plan': {
-      // Phase 1: Quick Plan (legacy/backward-compat when research is skipped)
-      console.log(JSON.stringify({
-        continue: true,
-        stopReason: '[MPL] Phase 1: Quick Plan in progress. Complete planning and HITL before proceeding.'
-      }));
-      break;
-    }
-
-    case 'phase1a-research': {
-      // Phase 1-A: Deep Research — check research.status for auto-transition
-      const research = state.research || {};
-
-      // BUG-6 fix: handle research error state to prevent infinite loop
-      if (research.error) {
-        // codex r7 on PR #222: phase1b-plan is exempt from the Phase 0
-        // artifact guard (it's the planning phase that produces
-        // decomposition + contracts). Transition unconditionally.
-        writeState(cwd, { current_phase: 'phase1b-plan', research: { status: 'skipped' } });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Research failed: ${research.error}. Skipping to Phase 1-B: Plan Generation (without research).`
-        }));
-        break;
-      }
-
-      if (research.status === 'completed' || research.status === 'skipped') {
-        // Research done → transition to Phase 1-B: Plan Generation
-        // (phase1b-plan exempt; see comment above)
-        writeState(cwd, { current_phase: 'phase1b-plan' });
-        const msg = research.status === 'skipped'
-          ? '[MPL] Research skipped. Transitioning to Phase 1-B: Plan Generation.'
-          : `[MPL] Research completed (${research.stages_completed?.length || 0} stages, ${research.findings_count || 0} findings, ${research.sources_count || 0} sources). Transitioning to Phase 1-B: Plan Generation.`;
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: msg
-        }));
-      } else {
-        // Research in progress — guide orchestrator
-        const currentStage = research.status || 'not started';
-        const stagesCompleted = research.stages_completed?.length || 0;
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Phase 1-A: Deep Research in progress (stage: ${currentStage}, ${stagesCompleted}/3 stages completed). Complete all research stages or skip to proceed.`
-        }));
-      }
-      break;
-    }
-
-    case 'phase1b-plan': {
-      // Phase 1-B: Plan Generation — same as phase1-plan but after research
-      // Orchestrator handles PLAN.md creation + HITL; no auto-transition
-      const reportPath = state.research?.report_path;
-      const reportNote = reportPath ? ` Research report: ${reportPath}.` : '';
-      console.log(JSON.stringify({
-        continue: true,
-        stopReason: `[MPL] Phase 1-B: Plan Generation in progress.${reportNote} Use research findings as input for planning agents. Complete PLAN.md and HITL before proceeding.`
-      }));
-      break;
-    }
-
-    case 'phase2-sprint': {
-      // v0.17 (#55): interview_depth guard removed. Phase 0 no longer has
-      // light/full dual-track — Stage 1 always runs full depth.
-
-      // Stage A Phase 1.6b: lazy-initialize state.release.current_cut_id
-      // on first sprint entry when goal_contract.mvp_scope is declared.
-      // This is the single lifecycle write point per RFC §4.5: subsequent
-      // advancement happens at release-finalize exit. Without an mvp_scope,
-      // current_cut_id stays null and the release path is never entered
-      // (existing pipeline behavior preserved).
-      //
-      // RFC §4.5: "Never re-entered for the same cut_id within a single
-      // pipeline run." A cohort already in `completed_cut_ids` MUST NOT be
-      // re-set here, even if `current_cut_id` is null and the contract
-      // still carries `mvp_scope`. This guards the phase3-gate → phase4-fix
-      // → recompose → phase2-sprint loop from spuriously re-running the
-      // mvp release path after the artifact has shipped.
-      if (state.release?.current_cut_id == null) {
-        const already = Array.isArray(state.release?.completed_cut_ids)
-          ? state.release.completed_cut_ids
-          : [];
-        if (!already.includes('mvp')) {
-          const gc = readGoalContract(cwd);
-          if (gc.exists && gc.contract?.mvp_scope) {
-            writeState(cwd, {
-              release: { ...(state.release || {}), current_cut_id: 'mvp' },
-            });
-            // Re-read so the routing below sees the freshly-set cohort.
-            state = readState(cwd);
-          }
-        }
-      }
-
-      // Check PLAN.md completion
-      const planStatus = checkPlanStatus(cwd);
-      if (!planStatus || planStatus.total === 0) {
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL] Phase 2: Sprint in progress. No PLAN.md found or no TODOs defined.'
-        }));
-        break;
-      }
-
-      const { total, completed, failed, deferred = 0 } = planStatus;
-      // #241 B6: `[~]` (deferred by decision) counts toward "no remaining
-      // work" for routing so the operator does not have to forge `[x]`
-      // on a deliberately-deferred task. deferred is surfaced in the
-      // stopReason so the operator still sees it.
-      const remaining = total - completed - failed - deferred;
-      const deferredTail = deferred > 0 ? `, ${deferred} deferred` : '';
-
-      if (remaining === 0) {
-        // Stage A Phase 1.6b: route based on `state.release.current_cut_id`.
-        // - non-null → cohort active, transition to release-gate (scoped Hard
-        //   1/2/3 will run there once Phase 1.6c lands; for now the stub
-        //   immediately advances to release-finalize)
-        // - null → whole-pipeline complete, transition to phase3-gate
-        //   (existing pre-Stage-A behavior, unchanged for projects without
-        //   mvp_scope)
-        const cohort = state.release?.current_cut_id;
-
-        // RFC §10 D-Q6 + Phase 1.4b: a cohort with FAILED TODOs MUST NOT be
-        // recorded as completed. release-finalize would otherwise append the
-        // cohort to `completed_cut_ids` (PR #185 codex review), flipping D-Q6
-        // immutability on for a cohort that never actually shipped its release
-        // artifact. The release path is reserved for cohorts whose sprint
-        // completed cleanly.
-        //
-        // Codex re-review caught a follow-up: routing to phase3-gate here
-        // would let an existing all-PASS state.gate_results path through to
-        // phase5-finalize while the cohort is still active — whole-pipeline
-        // finalize would start before the release path completes. Stay in
-        // phase2-sprint instead until the FAILED TODOs are cleared in PLAN.md
-        // (user/agent flips [FAILED] → [x] or removes the TODO). The release
-        // path resumes naturally when the sprint completes cleanly.
-        if (cohort && failed > 0) {
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `[MPL] Sprint resolved with FAILED TODOs (${completed} completed, ${failed} failed). ` +
-              `Active release cohort "${cohort}" cannot enter the release path while failures are present, ` +
-              `and routing to whole-pipeline phase3-gate would risk advancing past release-finalize when prior ` +
-              `gate evidence is all-PASS. Staying in phase2-sprint. ` +
-              `Clear the FAILED TODOs in PLAN.md (fix the failing tasks, then flip [FAILED] → [x]) to resume the release path.`
-          }));
-          break;
-        }
-
-        const nextPhase = cohort ? 'release-gate' : 'phase3-gate';
-        const target = cohort ? `release-gate(${cohort})` : 'Phase 3: Quality Gate';
-        // codex r2 on PR #222: variable nextPhase still needs the
-        // Phase 0 artifact guard — both branches land in protected phases.
-        if (emitPhase0BlockIfNeeded(cwd, nextPhase)) return;
-        writeState(cwd, { current_phase: nextPhase });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] All TODOs resolved (${completed} completed, ${failed} failed${deferredTail}). Transitioning to ${target}.`
-        }));
-      } else {
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Phase 2: Sprint in progress. ${completed}/${total} TODOs completed, ${failed} failed${deferredTail}, ${remaining} remaining.`
-        }));
-      }
-      break;
-    }
-
-    // === Stage A Release Path (Phase 1.6b stubs; 1.6c will add scoped Hard
-    // 1/2/3 and release-manifest / artifact creation) ===
-
-    case 'release-gate': {
-      // Phase 1.6c-i: scoped Hard 1/2/3 evidence routing.
-      //
-      // Reads `state.release.gate_results` (parallel subtree to top-level
-      // `state.gate_results`, per RFC §5.5 — scoped release evidence MUST
-      // NOT mix into whole-pipeline gates reserved for the final
-      // phase3-gate). mpl-gate-recorder populates this subtree only when
-      // `current_phase == 'release-gate'`; this handler consumes and
-      // routes. Evidence *production* (running scoped commands, deciding
-      // affected-tests scope for Hard 2, etc.) is gate-recorder's
-      // responsibility and lands separately.
-      //
-      // Routing per RFC §5.3.1:
-      //   - PASS → release-finalize
-      //   - FAIL → increment state.release.fix_loop_count, route back to
-      //     phase2-sprint, PRESERVE current_cut_id. Reset scoped
-      //     evidence so the next attempt starts clean.
-      //   - FAIL at threshold (count reaches state.release.max_fix_loops,
-      //     default 3) → circuit-break: pin cohort, stay at release-gate,
-      //     surface user-actionable message. User intervention required.
-      //   - MISSING → continue with stopReason guiding orchestrator to
-      //     produce evidence; same UX as phase3-gate partial/empty paths.
-      const cohort = state.release?.current_cut_id;
-      if (!cohort) {
-        // Defensive: release-gate without an active cohort is a state
-        // corruption. Surface and revert to phase3-gate (whole-pipeline).
-                if (emitPhase0BlockIfNeeded(cwd, 'phase3-gate')) return;
-                writeState(cwd, { current_phase: 'phase3-gate' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL] ⚠ release-gate entered with no active cohort. Reverting to phase3-gate.'
-        }));
-        break;
-      }
-
-      // The workspace `missing_gate_evidence` rule only controls MISSING
-      // message wording (in progress vs ⛔ BLOCKED) for parity with
-      // phase3-gate UX. The strict toggle for checkGateResults is ALWAYS
-      // true on release-gate, regardless of policy.
-      //
-      // Rationale (PR #186 codex/claude High #2): `state.release.gate_results`
-      // is a v6 subtree with no historical legacy boolean evidence to
-      // honor. The transitional zero-structured legacy fallback was added
-      // for top-level state.gate_results to ease the AD-0006 migration; on
-      // a brand-new subtree it would let `release.gate_results.hard1_passed
-      // = true` alone trigger PASS → release-finalize, bypassing the
-      // structured-only contract Phase 1.6c-i is meant to enforce.
-      const releaseGateRuleAction = resolveRuleAction(cwd, state, 'missing_gate_evidence');
-      const releaseMissingMessagingStrict = releaseGateRuleAction === 'block';
-      // Adapter: checkGateResults reads `state.gate_results`. Wrap the
-      // release subtree to reuse the structured-vs-legacy decision tree
-      // without duplicating logic across two gate paths. `strict: true`
-      // is hard-coded (see rationale above) — release-gate ALWAYS requires
-      // structured exits.
-      const releaseGates = state.release?.gate_results || {};
-      const releaseResults = checkGateResults({ gate_results: releaseGates }, { strict: true });
-
-      if (releaseResults.allPassed) {
-                if (emitPhase0BlockIfNeeded(cwd, 'release-finalize')) return;
-                writeState(cwd, { current_phase: 'release-finalize' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] release-gate(${cohort}): scoped Hard 1/2/3 passed (source=${releaseResults.source}). Transitioning to release-finalize.`
-        }));
-        break;
-      }
-
-      if (releaseResults.anyFailed) {
-        const existingRelease = state.release || {};
-        const currentReleaseFix = typeof existingRelease.fix_loop_count === 'number'
-          ? existingRelease.fix_loop_count
-          : 0;
-        const releaseMax = typeof existingRelease.max_fix_loops === 'number'
-          ? existingRelease.max_fix_loops
-          : 3;
-        // Avoid runaway increment when already pinned at threshold —
-        // repeated Stop ticks would otherwise grow the counter past the
-        // cap (cosmetic but noisy in /mpl-status).
-        const nextReleaseFix = currentReleaseFix >= releaseMax
-          ? currentReleaseFix
-          : currentReleaseFix + 1;
-        const willCircuitBreak = nextReleaseFix >= releaseMax;
-
-        if (willCircuitBreak) {
-          // Pin cohort, stay at release-gate. Only write when the count
-          // actually changed (skip writeState noise on repeat pinned ticks).
-          if (nextReleaseFix !== currentReleaseFix) {
-            writeState(cwd, {
-              release: { ...existingRelease, fix_loop_count: nextReleaseFix },
-            });
-          }
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `[MPL] ⛔ release-gate(${cohort}) circuit-break: release-scoped fix loop ${nextReleaseFix}/${releaseMax} (RFC §5.3.1). ` +
-              `Gate results: H1=${releaseResults.details.hard1}, H2=${releaseResults.details.hard2}, H3=${releaseResults.details.hard3}. ` +
-              `Cohort pinned. User intervention required: fix the failing tasks and reset state.release.fix_loop_count via mpl_state_write, ` +
-              `OR remove the cohort from goal-contract.yaml mvp_scope to abort the release path.`
-          }));
-          break;
-        }
-
-        // Budget remaining → route back to phase2-sprint. Reset scoped
-        // evidence on retry so mpl-gate-recorder writes fresh exits on
-        // the next attempt (same pattern as phase3-gate FAIL path).
-        // codex r3 on PR #222: protect this composite writeState too.
-        if (emitPhase0BlockIfNeeded(cwd, 'phase2-sprint')) return;
-        writeState(cwd, {
-          current_phase: 'phase2-sprint',
-          release: {
-            ...existingRelease,
-            fix_loop_count: nextReleaseFix,
-            gate_results: {
-              hard1_passed: null, hard2_passed: null, hard3_passed: null,
-              hard1_baseline: null, hard2_coverage: null, hard3_resilience: null,
-            },
-          },
-        });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] release-gate(${cohort}) FAILED (source=${releaseResults.source}). ` +
-            `H1=${releaseResults.details.hard1}, H2=${releaseResults.details.hard2}, H3=${releaseResults.details.hard3}. ` +
-            `Scoped fix loop ${nextReleaseFix}/${releaseMax}. Returning to phase2-sprint; cohort "${cohort}" preserved. ` +
-            `Fix the failing tasks, then re-run scoped Hard 1/2/3 — mpl-gate-recorder will repopulate state.release.gate_results.`
-        }));
-        break;
-      }
-
-      // MISSING (zero or partial structured evidence). No transition;
-      // surface guidance and let the orchestrator produce evidence. The
-      // workspace strict policy changes the message wording only — the
-      // structured-only contract is always enforced (see above).
-      //
-      // Note: `missingEvidence` is guaranteed non-empty here because
-      // checkGateResults returns the MISSING branch only when at least
-      // one of hard{1,2,3} is absent. The full-PASS path returns earlier.
-      const missing = releaseResults.missingEvidence;
-      const verb = releaseMissingMessagingStrict ? '⛔ BLOCKED' : 'in progress';
-      const tail = releaseMissingMessagingStrict
-        ? 'Strict enforcement requires all 3 scoped gates to be recorded by mpl-gate-recorder via real Bash exit codes. Self-reported booleans are not accepted.'
-        : `Run the missing scoped gates so mpl-gate-recorder writes structured evidence into state.release.gate_results; the loop will continue once all 3 are recorded.`;
-      console.log(JSON.stringify({
-        continue: true,
-        stopReason: `[MPL] release-gate(${cohort}) ${verb}: missing scoped Hard 1/2/3 evidence (${missing.join(', ')}). ${tail}`
-      }));
-      break;
-    }
-
-    case 'release-finalize': {
-      // Phase 1.6c-ii: write release-manifest.json + evidence-summary.md +
-      // gate-results.json under .mpl/mpl/releases/{cut_id}/ before the
-      // existing append+clear step. RFC §5.4 requires the manifest write
-      // to precede `completed_cut_ids` append — a cohort is only "released"
-      // (and D-Q6 immutable) once its manifest is shipped.
-      //
-      // Snapshot identifiers (commit_sha/tree_sha/snapshot_ref) are
-      // placeholders here; 1.6c-iii populates them via git rev-parse /
-      // update-ref. Optional user-visible artifact creation (draft_pr /
-      // branch / tag) also lands in 1.6c-iii. The manifest schema is
-      // stable across the 1.6c-ii → 1.6c-iii boundary so the upcoming
-      // diff is small.
-      //
-      // RFC §5.5 invariants preserved: this handler MUST NOT set
-      // `state.finalize_done=true` and MUST NOT transition `current_phase`
-      // to `completed`. Both remain exclusive to phase5-finalize.
-      //
-      // Codex r5 on PR #222 [data-integrity]: Phase 0 artifact guard
-      // MUST run at the START of the release-finalize case, BEFORE
-      // createSnapshotRef / attemptArtifactCreation / atomicWriteFile
-      // create artifacts on disk. Otherwise a missing-artifacts state
-      // would leave shipped-looking manifest/ref files while
-      // state.release.completed_cut_ids is not appended — release
-      // immutability drift.
-      if (emitPhase0BlockIfNeeded(cwd, 'release-finalize')) return;
-      const cur = state.release?.current_cut_id;
-      if (!cur) {
-                if (emitPhase0BlockIfNeeded(cwd, 'phase3-gate')) return;
-                writeState(cwd, { current_phase: 'phase3-gate' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL] ⚠ release-finalize entered with no active cohort. Routing to phase3-gate.'
-        }));
-        break;
-      }
-
-      // Read contract + decomposition for the manifest builder. Both are
-      // required: contract supplies goal_trace (AC/AX), decomposition
-      // supplies phases per cut. Missing either is an actionable error —
-      // refuse to write a degraded manifest and stay at release-finalize.
-      //
-      // PR #187 round-2 codex+claude High: an INVALID contract (e.g.,
-      // `mvp_scope: { artifact: draft_pr }` with no AC/AX) was previously
-      // passed through because the handler only checked `gcRead.exists`.
-      // The validator already flags `mvp_scope.acceptance_criteria_or_variation_axes`
-      // and other structural failures; route any `!gcRead.valid` through
-      // the same bail mechanism so a degraded "released" manifest with
-      // empty goal_trace cannot flip D-Q6 immutability. Library-layer
-      // defense (resolveCutDescriptor rejecting empty AC+AX) is the
-      // backstop for callers that bypass `readGoalContract`.
-      const gcRead = readGoalContract(cwd);
-      if (gcRead.exists && !gcRead.valid) {
-        const reasonList = Array.isArray(gcRead.missing) && gcRead.missing.length > 0
-          ? gcRead.missing.join(', ')
-          : '(no detail)';
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] ⛔ release-finalize(${cur}): goal-contract is invalid (${reasonList}). ` +
-            `Cohort NOT appended to completed_cut_ids. Fix .mpl/goal-contract.yaml — the validator at ` +
-            `mpl-goal-contract.validateGoalContractText lists what is required. ` +
-            `Staying at release-finalize until resolved.`
-        }));
-        break;
-      }
-      const contract = gcRead.exists ? gcRead.contract : null;
-      let graph = null;
-      let decomposition = null;
-      try {
-        const decompPath = join(cwd, '.mpl', 'mpl', 'decomposition.yaml');
-        if (existsSync(decompPath)) {
-          const decompText = readFileSync(decompPath, 'utf-8');
-          graph = parsePhaseContractGraphText(decompText);
-          decomposition = parseDecompositionGoalTraceText(decompText);
-        }
-      } catch {
-        graph = null;
-        decomposition = null;
-      }
-
-      const writtenAt = new Date().toISOString();
-      const manifest = buildReleaseManifest({ cutId: cur, state, contract, graph, decomposition, now: writtenAt });
-      if (!manifest) {
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] ⛔ release-finalize(${cur}): cohort descriptor missing from contract/decomposition. ` +
-            `Cannot build release manifest. Verify .mpl/goal-contract.yaml mvp_scope and ` +
-            `.mpl/mpl/decomposition.yaml mvp.phases are present and match the active cut_id. ` +
-            `Staying at release-finalize until resolved.`
-        }));
-        break;
-      }
-
-      // Phase 1.6c-iii: snapshot ref creation MUST succeed before any
-      // file is written or the cohort is marked released. RFC §5.4 §232
-      // explicitly requires "gate PASS + release-manifest write + snapshot
-      // ref creation all succeed" before appending completed_cut_ids;
-      // the snapshot ref is the immutability anchor (the commit that the
-      // shipped manifest claims to pin). PR #188 codex review: prior
-      // implementation flipped through to release on snapshot failure
-      // with null commit_sha/tree_sha/snapshot_ref, which would let
-      // D-Q6 immutability turn on for a manifest that pins nothing.
-      const snapshot = createSnapshotRef(cwd, cur);
-      if (!snapshot.ok) {
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] ⛔ release-finalize(${cur}): snapshot ref creation failed (${snapshot.reason}). ` +
-            `Cohort NOT appended to completed_cut_ids per RFC §5.4 (snapshot ref is the immutability anchor). ` +
-            `Resolve the underlying git issue (e.g., not a git repo, detached HEAD) and let the loop retry.`
-        }));
-        break;
-      }
-      manifest.commit_sha = snapshot.commit_sha;
-      manifest.tree_sha = snapshot.tree_sha;
-      manifest.snapshot_ref = snapshot.snapshot_ref;
-
-      // Phase 1.6c-iii: attempt user-visible artifact (tag / branch /
-      // draft_pr) BEFORE writing the manifest so the failure result is
-      // captured in the single canonical manifest write — no re-write,
-      // no disk ↔ stopReason divergence (PR #188 claude review #2).
-      // The attempt is best-effort per RFC §5.4: failures are recorded
-      // in the manifest (artifact_creation_failed) and surfaced to the
-      // user, but they do NOT block `completed_cut_ids` append because
-      // artifact creation depends on external tools (gh CLI, remote
-      // permissions) that are out of MPL's control. Snapshot ref is
-      // distinct from this — it's the immutability anchor and was
-      // already enforced above.
-      let artifactAttempt = null;
-      if (manifest.artifact && manifest.artifact !== 'release_manifest') {
-        artifactAttempt = attemptArtifactCreation({
-          cwd,
-          cutId: cur,
-          artifact: manifest.artifact,
-          commitSha: snapshot.commit_sha,
-          snapshotRef: snapshot.snapshot_ref,
-        });
-        if (artifactAttempt.artifact_creation_failed) {
-          manifest.artifact_creation_failed = artifactAttempt.artifact_creation_failed;
-        }
-      }
-
-      const evidence = buildEvidenceSummary({ cutId: cur, state, contract, graph, decomposition, now: writtenAt });
-      const gateSnapshot = buildGateResultsSnapshot(state, writtenAt);
-
-      const releaseDir = join(cwd, RELEASE_DIR_REL_PATH, cur);
-      const manifestPath = join(releaseDir, 'release-manifest.json');
-      try {
-        mkdirSync(releaseDir, { recursive: true });
-        // 0o644 (not the 0o600 used for state.json): release artifacts
-        // are designed for consumption by CI runners, mpl-status, and
-        // future tooling that may run as a different uid on a shared
-        // dev box / CI agent (claude review #1 on PR #187). state.json
-        // stays 0o600 because it carries session-internal evidence;
-        // these files are the shippable record of the release.
-        //
-        // Atomic temp+rename (claude review #2 on PR #187) so a mid-write
-        // failure leaves the on-disk state at the prior generation
-        // instead of an asymmetric "manifest written but summary
-        // missing" state.
-        atomicWriteFile(
-          manifestPath,
-          JSON.stringify(manifest, null, 2) + '\n',
-          0o644
-        );
-        atomicWriteFile(
-          join(releaseDir, 'evidence-summary.md'),
-          evidence.endsWith('\n') ? evidence : evidence + '\n',
-          0o644
-        );
-        atomicWriteFile(
-          join(releaseDir, 'gate-results.json'),
-          JSON.stringify(gateSnapshot, null, 2) + '\n',
-          0o644
-        );
-      } catch (err) {
-        // Manifest write failure must NOT advance the lifecycle (RFC §5.4:
-        // append only after manifest write succeeds). Surface and pin.
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] ⛔ release-finalize(${cur}): manifest write failed (${err?.message || 'unknown'}). ` +
-            `Cohort NOT appended to completed_cut_ids. Resolve the filesystem error under ` +
-            `${RELEASE_DIR_REL_PATH}/${cur}/ and let the loop retry.`
-        }));
-        break;
-      }
-
-      const existing = state.release || {};
-      const already = Array.isArray(existing.completed_cut_ids) ? existing.completed_cut_ids : [];
-      const completed = already.includes(cur) ? already : [...already, cur];
-
-      // Stage A simplification: decomposer (Phase 1.2 / PR #182) emits
-      // `release_cuts: []` so there is never a "next cut" to advance to.
-      // The orchestrator routes directly to whole-pipeline phase3-gate.
-      // Multi-cohort cut chaining (auto-proposed extension cuts) is RFC
-      // §10 D-Q2 Stage B work.
-      const nextCutId = null;
-      const nextPhase = 'phase3-gate';
-
-      // codex r5/r6 on PR #222: the release-finalize case is guarded at
-      // its TOP (line 740-ish) before any release artifact write. The
-      // r2 post-write guard here was redundant AND harmful — if Phase 0
-      // files were removed between the entry guard and this point, we'd
-      // skip the state update while leaving manifest/snapshot artifacts
-      // on disk (release immutability drift). Drop the post-write guard
-      // and rely on the single entry guard for this case.
-      writeState(cwd, {
-        release: {
-          ...existing,
-          completed_cut_ids: completed,
-          current_cut_id: nextCutId,
-          fix_loop_count: 0,
-        },
-        current_phase: nextPhase,
-      });
-      // Build a status-aware tail so users immediately see snapshot +
-      // artifact outcome at the single Stop hook surface. Snapshot is
-      // always ok here (failure path bails earlier); push may be soft-
-      // failed so surface that explicitly.
-      const snapshotTail = snapshot.pushed
-        ? `snapshot ${snapshot.snapshot_ref} @ ${snapshot.commit_sha.slice(0, 12)} (pushed)`
-        : `snapshot ${snapshot.snapshot_ref} @ ${snapshot.commit_sha.slice(0, 12)} (local only, push: ${snapshot.push_reason})`;
-      const artifactTail = (() => {
-        if (!manifest.artifact || manifest.artifact === 'release_manifest') return 'artifact=release_manifest (no external push)';
-        if (!artifactAttempt) return `artifact=${manifest.artifact} skipped`;
-        if (artifactAttempt.artifact_creation_failed) {
-          return `⚠ artifact=${manifest.artifact} FAILED (${artifactAttempt.artifact_creation_failed.reason})`;
-        }
-        const noopSuffix = artifactAttempt.result?.noop ? ' (noop — already at this commit)' : '';
-        return `artifact=${manifest.artifact} created${noopSuffix}`;
-      })();
-      console.log(JSON.stringify({
-        continue: true,
-        stopReason: `[MPL] release-finalize(${cur}): cohort completed. Manifest written to ${RELEASE_DIR_REL_PATH}/${cur}/. ` +
-          `${snapshotTail}. ${artifactTail}. ` +
-          `Stage A single-cohort path: transitioning to whole-pipeline phase3-gate.`
-      }));
-      break;
-    }
-
-    case 'phase3-gate': {
-      // #241 B2 (delivered via #248): the previous defense-in-depth
-      // FORCE-REVERTED to phase2-sprint when `state.release.current_cut_id`
-      // was non-null. That over-blocked the legitimate case where the
-      // operator has PASS gate evidence but `current_cut_id` is stale
-      // (e.g., from a partially-rolled-back cohort, hand-edited state,
-      // or recompose-driven re-entry).
-      //
-      // New behavior: emit an advisory stopReason recommending
-      // `mpl-recover` to clear the stale cohort, but DO NOT auto-revert
-      // `state.current_phase`. The existing gate evidence check below
-      // remains the actual gate — if gate_results are PASS, the phase
-      // transitions to phase5-finalize normally; if gates haven't been
-      // recorded, the missing-evidence branches still block.
-      //
-      // Operator workflow:
-      //   - cohort is genuinely active → run `mpl-recover` to clear
-      //     `current_cut_id`, or complete release-gate → release-finalize.
-      //   - cohort is stale → either accept the advisory and proceed
-      //     (gate evidence still gates), or run `mpl-recover` to clean
-      //     up the state.
-      const cohortAdvisory = state.release?.current_cut_id
-        ? `[MPL #241 B2 / #248] ⚠ Advisory: phase3-gate entered while release cohort "${state.release.current_cut_id}" is still tracked in state.release.current_cut_id. ` +
-          `If no release path is in progress, run \`mpl-recover\` to clear the stale cohort marker; if the cohort IS active, complete release-gate → release-finalize before whole-pipeline finalize. ` +
-          `(Per #248: not auto-reverting state.current_phase; gate evidence below remains the actual gate.) `
-        : '';
-
-      // Per-rule policy (P0-2, #110): `missing_gate_evidence` resolves the
-      // strict toggle for checkGateResults. Default is 'warn' per #110 §정책
-      // (transitional — surface only, no block). Workspace can opt-in to
-      // 'block' to halt phase3-gate transitions until mpl-gate-recorder writes
-      // structured exits, or 'off' to suppress the legacy fallback ⚠ entirely.
-      // Precedence: state.enforcement > .mpl/config.json > plugin baseline.
-      const gateRuleAction = resolveRuleAction(cwd, state, 'missing_gate_evidence');
-      const enforcementStrict = gateRuleAction === 'block';
-      const gateResults = checkGateResults(state, { strict: enforcementStrict });
-
-      const fallbackWarn = (gateResults.source === 'legacy' && gateRuleAction !== 'off')
-        ? ' ⚠ Using legacy gate boolean fallback (no structured evidence in state.gate_results.hard{1,2,3}_{baseline,coverage,resilience}). exp16 strict mode will block this transition. Run real verification commands so mpl-gate-recorder can record exit codes.'
-        : '';
-
-      if (gateResults.allPassed) {
-        // #241 B2 (#248) codex r1 fix: a genuinely active release cohort
-        // (current_cut_id set) with all-PASS top-level gates MUST NOT
-        // bypass release-gate / release-finalize. The whole-pipeline
-        // finalize path would skip completed_cut_ids accounting, the
-        // release manifest, and the cohort-scoped evidence files. Route
-        // to release-gate instead — the cohort's release path validates
-        // and finalizes the cut; if the marker is stale, the operator
-        // clears it via mpl-recover (per the advisory) and re-enters
-        // phase3-gate. Both paths converge cleanly.
-        if (state.release?.current_cut_id) {
-          if (emitPhase0BlockIfNeeded(cwd, 'release-gate')) return;
-          writeState(cwd, { current_phase: 'release-gate' });
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `${cohortAdvisory}[MPL] All Quality Gates passed (source=${gateResults.source}). Active cohort "${state.release.current_cut_id}" detected — routing to release-gate (not whole-pipeline phase5-finalize) to complete the cohort's release path. Run \`mpl-recover\` first if the cohort marker is stale.${fallbackWarn}`
-          }));
-          break;
-        }
-                if (emitPhase0BlockIfNeeded(cwd, 'phase5-finalize')) return;
-                writeState(cwd, { current_phase: 'phase5-finalize' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `${cohortAdvisory}[MPL] All Quality Gates passed (source=${gateResults.source}). Transitioning to Phase 5: Finalize.${fallbackWarn}`
-        }));
-      } else if (gateResults.anyFailed) {
-        // Gate failed → Phase 4
-        // Preserve existing fix_loop_count to prevent infinite loop bypass
-        // (only initialize to 0 on first entry, not on re-entry from Phase 3)
-        const currentFixCount = state.fix_loop_count || 0;
-        // codex r3 on PR #222: composite writeState into a protected phase
-        // needs the Phase 0 artifact guard too — without it, a gate-failure
-        // transition would not only advance into phase4-fix unprotected
-        // but also nuke the recorded failure evidence (the gate_results
-        // reset below).
-        if (emitPhase0BlockIfNeeded(cwd, 'phase4-fix')) return;
-        // Reset both legacy and structured gate evidence on retry to prevent stale data.
-        writeState(cwd, {
-          current_phase: 'phase4-fix',
-          fix_loop_count: currentFixCount,
-          gate_results: {
-            hard1_passed: null, hard2_passed: null, hard3_passed: null,
-            hard1_baseline: null, hard2_coverage: null, hard3_resilience: null,
-          }
-        });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `${cohortAdvisory}[MPL] Quality Gate failed (source=${gateResults.source}). Gate results: H1=${gateResults.details.hard1}, H2=${gateResults.details.hard2}, H3=${gateResults.details.hard3}. Transitioning to Phase 4: Fix Loop.${fallbackWarn}`
-        }));
-      } else if (gateResults.source === 'structured' && gateResults.missingEvidence.length > 0) {
-        // Partial structured evidence: gate-recorder produced some entries but not all.
-        // Issue #102 spec ("missing → false") + non-strict UX (E-gap from PR #119 smoke):
-        // surface the explicit missing list whether or not strict mode is on. The
-        // wording adapts to whether transition is hard-blocked (strict) or "in progress"
-        // (non-strict — phase will retry on next Stop event after the missing gates record).
-        const verb = enforcementStrict ? '⛔ BLOCKED' : 'in progress';
-        const tail = enforcementStrict
-          ? 'Strict enforcement requires all 3 gates to be recorded by mpl-gate-recorder via real Bash exit codes. Self-reported booleans are not accepted.'
-          : `Run the missing gates so mpl-gate-recorder writes structured evidence; the loop will continue once all 3 ${gateResults.missingEvidence.length === 3 ? 'are' : 'remaining are'} recorded.`;
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `${cohortAdvisory}[MPL] Phase 3 ${verb}: missing structured gate evidence (${gateResults.missingEvidence.join(', ')}). ${tail}`
-        }));
-      } else {
-        // No gate evidence at all (zero structured + zero legacy).
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `${cohortAdvisory}[MPL] Phase 3: Quality Gate in progress. Run all 3 gates before proceeding.${fallbackWarn}`
-        }));
-      }
-      break;
-    }
-
-    case 'phase4-fix': {
-      const fixCount = state.fix_loop_count || 0;
-      const maxFix = state.max_fix_loops || 10;
-
-      if (fixCount >= maxFix) {
-        // Fix loop limit reached → Phase 5 (partial completion)
-                if (emitPhase0BlockIfNeeded(cwd, 'phase5-finalize')) return;
-                writeState(cwd, { current_phase: 'phase5-finalize' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL] Fix loop limit reached (${fixCount}/${maxFix}). Transitioning to Phase 5: Finalize (partial completion).`
-        }));
-      } else {
-        // H1: Check convergence before continuing
-        const convergenceResult = checkConvergence(state);
-        // #241 B3 (delivered via #248): the previous behavior
-        // auto-finalized on the FIRST detected stagnation tick. A
-        // brief mid-bug plateau (e.g., one tick of zero improvement
-        // while the agent reads files) was enough to force partial
-        // completion. New behavior:
-        //   - Count consecutive stagnant/regressing ticks in
-        //     `state.fix_loop.stagnation_tick_count`.
-        //   - Each tick where status ∈ {'stagnating', 'regressing'}
-        //     increments the counter; otherwise reset to 0.
-        //   - Auto-finalize fires ONLY when:
-        //       count >= `convergence.stagnation_window` (default 3)
-        //       AND `.mpl/config.json:convergence.auto_finalize_on_stagnation`
-        //         is true (default false).
-        //   - Until both conditions hold, emit an advisory stopReason
-        //     and stay in the fix loop. The operator can intervene
-        //     manually (force-finalize, abort, etc.) or let the loop
-        //     re-converge.
-        const cfg = loadConfig(cwd);
-        const convergenceCfg = cfg.convergence || {};
-        const stagnationWindow = Number.isInteger(convergenceCfg.stagnation_window)
-          && convergenceCfg.stagnation_window > 0
-          ? convergenceCfg.stagnation_window
-          : 3;
-        const autoFinalizeOnStagnation = convergenceCfg.auto_finalize_on_stagnation === true;
-
-        if (convergenceResult.status === 'stagnating' || convergenceResult.status === 'regressing') {
-          const prevCount = state.fix_loop?.stagnation_tick_count || 0;
-          const newCount = prevCount + 1;
-          if (newCount >= stagnationWindow && autoFinalizeOnStagnation) {
-            if (emitPhase0BlockIfNeeded(cwd, 'phase5-finalize')) return;
-            writeState(cwd, {
-              current_phase: 'phase5-finalize',
-              fix_loop: { ...(state.fix_loop || {}), stagnation_tick_count: newCount },
-            });
-            console.log(JSON.stringify({
-              continue: true,
-              stopReason: `[MPL] Convergence ${convergenceResult.status} for ${newCount}/${stagnationWindow} consecutive ticks ` +
-                `(delta: ${convergenceResult.delta?.toFixed(3)}; auto_finalize_on_stagnation=true). ` +
-                `Transitioning to Phase 5: Finalize (partial completion).`
-            }));
-          } else {
-            // Advisory: increment counter, stay in fix loop. The next tick
-            // either resets (if convergence resumes) or escalates.
-            writeState(cwd, {
-              fix_loop: { ...(state.fix_loop || {}), stagnation_tick_count: newCount },
-            });
-            const autoFinalizeNote = autoFinalizeOnStagnation
-              ? ` After ${stagnationWindow} consecutive stagnant ticks, this will auto-finalize.`
-              : ` Auto-finalize is off (set .mpl/config.json:convergence.auto_finalize_on_stagnation to true to enable). Continue fixing, force-finalize manually, or abort.`;
-            console.log(JSON.stringify({
-              continue: true,
-              stopReason: `[MPL #241 B3 / #248] Advisory: convergence ${convergenceResult.status} on tick ${newCount}/${stagnationWindow} ` +
-                `(delta: ${convergenceResult.delta?.toFixed(3)}).${autoFinalizeNote}`
-            }));
-          }
-        } else {
-          // Convergence is fine — reset the stagnation counter.
-          if ((state.fix_loop?.stagnation_tick_count || 0) > 0) {
-            writeState(cwd, {
-              fix_loop: { ...(state.fix_loop || {}), stagnation_tick_count: 0 },
-            });
-          }
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `[MPL] Phase 4: Fix Loop ${fixCount}/${maxFix}. Continue fixing or re-run Quality Gate.`
-          }));
-        }
-      }
-      break;
-    }
-
-    case 'phase5-finalize': {
-      // Finalize: do NOT auto-transition to completed here.
-      // The orchestrator must complete finalization tasks (extract learnings, commit)
-      // and then manually set current_phase to 'completed' via writeState.
-      const finalized = state.finalize_done === true;
-      if (finalized) {
-                if (emitPhase0BlockIfNeeded(cwd, 'completed')) return;
-                writeState(cwd, { current_phase: 'completed' });
-        console.log(JSON.stringify({
-          continue: false,
-          stopReason: '[MPL] Phase 5: Finalize complete. MPL pipeline finished.'
-        }));
-      } else {
-        // Protocol load enforcement: remind orchestrator to load finalize protocol
-        // This prevents the exp5 failure where finalize rules were never read
-        const protocolReminder = '\n\nIMPORTANT: Before proceeding, you MUST read the finalize protocol documents:\n' +
-          '1. Read the gate execution protocol (mpl-run-execute-gates or equivalent)\n' +
-          '2. Read the finalize protocol (mpl-run-finalize or equivalent)\n' +
-          '3. Execute all Hard Gates (H1: Build+Lint+Type, H2: Full Test Suite, H3: Contract Diff Guard)\n' +
-          '4. Run project-root-level tests (cargo test --workspace, npx vitest run, pytest, etc.)\n' +
-          '5. Check platform-constraints.md violations if it exists in .mpl/mpl/phase0/';
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL] Phase 5: Finalize in progress. Extract learnings, commit, then set state.finalize_done = true to complete.' + protocolReminder
-        }));
-      }
-      break;
-    }
-
-    // === Small Pipeline Phases (3-Phase Lightweight) ===
-
-    case 'small-plan': {
-      // RFC §10 D-Q7: small-pipeline and `mvp_scope` are mutually exclusive.
-      // Small-pipeline answers "the whole task is small, run a lightweight
-      // pipeline." MVP cut answers "the task is large, ship a subset first."
-      // Conflating the two doubles the state-machine complexity for marginal
-      // gain. When mvp_scope is declared, the project MUST take the full MPL
-      // pipeline with the release path; small-pipeline entry is rejected
-      // here rather than silently downgrading the contract.
-      const gc = readGoalContract(cwd);
-      if (gc.exists && gc.contract?.mvp_scope) {
-        console.log(JSON.stringify({
-          continue: false,
-          decision: 'block',
-          reason: '[MPL] small-pipeline is not available when goal_contract.mvp_scope is declared. ' +
-            'Use the full MPL pipeline so the Stage A release path (release-gate → release-finalize) ' +
-            'can deliver the user-declared MVP cohort. ' +
-            'Either run the full pipeline (recommended) or remove mvp_scope from .mpl/goal-contract.yaml and re-enter small-plan.'
-        }));
-        break;
-      }
-
-      // Small Plan: orchestrator handles, no auto-transition
-      console.log(JSON.stringify({
-        continue: true,
-        stopReason: '[MPL-Small] Phase 1: Small Plan in progress. Complete planning and HITL before proceeding.'
-      }));
-      break;
-    }
-
-    case 'small-sprint': {
-      // D-Q7 guard intentionally fires only at `case 'small-plan'`. If a
-      // user lands here after entering small-plan with mvp_scope absent
-      // (legitimate path), the contract may have been edited mid-flow to
-      // add mvp_scope. Re-checking here is *not* supported in Phase 1.6a
-      // — the cost is replicating the guard in three places (-sprint,
-      // -verify) for an edge case where the user is intentionally
-      // restructuring mid-pipeline. Recommended workflow for that case:
-      // cancel the small pipeline and restart with the full pipeline.
-      // Check PLAN.md completion (reuse checkPlanStatus)
-      const smallPlanStatus = checkPlanStatus(cwd);
-      if (!smallPlanStatus || smallPlanStatus.total === 0) {
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL-Small] Phase 2: Sprint in progress. No PLAN.md found or no TODOs defined.'
-        }));
-        break;
-      }
-
-      const { total: sTotal, completed: sCompleted, failed: sFailed, deferred: sDeferred = 0 } = smallPlanStatus;
-      // #241 B6: `[~]` deferred counts toward "no remaining" for the
-      // small-sprint path too.
-      const sRemaining = sTotal - sCompleted - sFailed - sDeferred;
-      const sDeferredTail = sDeferred > 0 ? `, ${sDeferred} deferred` : '';
-
-      if (sRemaining === 0) {
-        // All TODOs resolved → small-verify
-        writeState(cwd, { current_phase: 'small-verify' });
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL-Small] All TODOs resolved (${sCompleted} completed, ${sFailed} failed${sDeferredTail}). Transitioning to Phase 3: Verify.`
-        }));
-      } else {
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: `[MPL-Small] Phase 2: Sprint in progress. ${sCompleted}/${sTotal} TODOs completed, ${sFailed} failed${sDeferredTail}, ${sRemaining} remaining.`
-        }));
-      }
-      break;
-    }
-
-    case 'small-verify': {
-      // Simplified verification: only hard2 (code review)
-      const smallGate = state.gate_results || {};
-
-      if (smallGate.hard2_passed === true) {
-        // Code review passed → completed.
-        // codex r2 on PR #222 [contract-break]: the small pipeline is a
-        // separate lightweight flow (small-plan → small-sprint →
-        // small-verify) that intentionally skips Phase 0 — it does NOT
-        // produce raw-scan.md / design-intent.yaml / contracts. Do NOT
-        // guard the small-pipeline completion against Phase 0 artifacts.
-        writeState(cwd, { current_phase: 'completed' });
-        console.log(JSON.stringify({
-          continue: false,
-          stopReason: '[MPL-Small] Verification passed. Pipeline complete. Extract learnings and commit.'
-        }));
-      } else if (smallGate.hard2_passed === false) {
-        const smallFixCount = state.fix_loop_count || 0;
-        const smallMaxFix = state.max_fix_loops || 3;
-
-        if (smallFixCount >= smallMaxFix) {
-          // Fix loop limit reached → completed (partial).
-          // codex r2 [contract-break]: small-pipeline exempt (see above).
-          writeState(cwd, { current_phase: 'completed' });
-          console.log(JSON.stringify({
-            continue: false,
-            stopReason: `[MPL-Small] Fix loop limit reached (${smallFixCount}/${smallMaxFix}). Completing with partial results. Extract learnings.`
-          }));
-        } else {
-          // Review failed, retries remaining → back to small-sprint
-          writeState(cwd, { current_phase: 'small-sprint', fix_loop_count: smallFixCount + 1 });
-          console.log(JSON.stringify({
-            continue: true,
-            stopReason: `[MPL-Small] Code review failed. Retry ${smallFixCount + 1}/${smallMaxFix}. Returning to Sprint for fixes.`
-          }));
-        }
-      } else {
-        // Gate not yet evaluated
-        console.log(JSON.stringify({
-          continue: true,
-          stopReason: '[MPL-Small] Phase 3: Verify in progress. Run code review before proceeding.'
-        }));
-      }
-      break;
-    }
-
-    default: {
-      // Unknown or completed phase
-      console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-    }
+  // Forward to the legacy main flow (heavy file-write paths).
+  try {
+    await legacy.runLegacyMain(input);
+  } catch {
+    emitContinueTrue();
   }
 }
 
-main().catch(() => {
-  console.log(JSON.stringify({ continue: true, suppressOutput: true }));
-});
-
-export { checkPlanStatus, checkGateResults };
+if (isMain) {
+  main().catch(() => emitContinueTrue());
+}
